@@ -7,10 +7,10 @@ Gem::Specification.new do |gem|
   gem.authors = ["Bill Katz", "Ian Terrell", "Jan Krutisch"]
   gem.files = Dir["{lib}/**/*.rb", "{config}/**/*"]
   gem.dependencies
-  gem.add_development_dependency 'rails', '~> 4.0'
+  gem.add_development_dependency 'rails', '< 6.0.0', '>= 4.0.0'
   gem.add_development_dependency 'sqlite3'
   gem.add_development_dependency "rspec-rails", "~> 3.3"
   gem.add_development_dependency "rdoc"
-  gem.add_runtime_dependency 'activesupport', '~> 4.0'
+  gem.add_runtime_dependency 'activesupport', '>= 4.0.0', '< 6.0.0'
   gem.version = "3.0.0"
 end
