@@ -1,6 +1,6 @@
 class DummyGammaController < ApplicationController
   permit "noone"
-  
+
   def index
   end
 
@@ -10,34 +10,34 @@ protected
   end
 
   def handle_require_user_redirection_for_html
-    render :text => nil, :status => :ok
+    head :ok
   end
-  
+
   def handle_permission_denied_redirection_for_html
-    render :text => nil, :status => :ok
+    head :ok
   end
 
   def handle_require_user_redirection_for_xml
-    render :text => nil, :status => :ok
+    head :ok
   end
-  
+
   def handle_permission_denied_redirection_for_xml
-    render :text => nil, :status => :ok
+    head :ok
   end
-  
+
   def handle_require_user_redirection_for_json
-    render :text => nil, :status => :ok
+    head :ok
   end
-  
+
   def handle_permission_denied_redirection_for_json
-    render :text => nil, :status => :ok
+    head :ok
   end
-  
+
   def handle_require_user_redirection_for_js
-    render :text => nil, :status => :ok
+    head :ok
   end
-  
+
   def handle_permission_denied_redirection_for_js
-    render :text => nil, :status => :ok
+    head :ok
   end
 end
